pyproject.toml
[tool.poetry]
name = "poetry-publish"
version = "0.4.1"
description = "Helper to build and upload a project that used poetry to PyPi, with prechecks"
authors = ["JensDiemer <git@jensdiemer.de>"]
keywords=['poetry']
classifiers = [
    # http://pypi.python.org/pypi?%3Aaction=list_classifiers
    'Development Status :: 5 - Production/Stable',
    'Environment :: Console',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: Implementation :: PyPy',
    'Operating System :: OS Independent',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Utilities',
]
include = ['AUTHORS', 'LICENSE']

# Will be generated from README.creole with: 'poetry run update_rst_readme'
readme='README.rst'

[tool.poetry.dependencies]
python = ">=3.6,<4.0.0"
python-creole = {version = "*", optional = true}
twine = "*"

[tool.poetry.dev-dependencies]
tox = "*"
pytest = "*"
pytest-cov = "*"
isort = "*"
flake8 = "*"
flynt = "*"
autopep8 = "*"
pyupgrade = "*"

[tool.poetry.extras]
creole = ["python-creole"]

[tool.poetry.scripts]
update_rst_readme = "poetry_publish.self:update_poetry_publish_readme"
publish = "poetry_publish.self:publish_poetry_publish"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
